tests/wpt/mozilla/tests/mozilla/bluetooth/getPrimaryServices/get-same-service.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.two_heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => Promise.all([
        gattServer.getPrimaryServices(heart_rate.name),
        gattServer.getPrimaryServices(heart_rate.name)]))
    .then(services => assert_array_equals(services[0], services[1]));
}, 'Calls to get the same service should return the same object.');
</script>
